fix(user): fixing response for login to return correct value for expires_at field
field was representing the timestamp but should be the time until
expiration in seconds

+++ src/model/user.rs
@@ -38,14 +38,14 @@ pub struct UserTokenDTO {
     pub access_token: String,
     // pub refresh_token: String,
     pub token_type: String,
-    pub expires_in: usize,
+    pub expires_in: u64,
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
 pub struct UserClaimsDTO {
     pub user_id: String,
     pub is_admin: bool,
-    pub exp: usize,
+    pub exp: u64,
     pub jti: String,
 }
